@@ -1,6 +1,6 @@
 {
 	"name": "archive-utility",
-	"version": "2.0.0",
+	"version": "2.0.1",
 	"description": "a Discord bot to archive text-channels",
 	"main": "dist/index.js",
 	"type": "module",
@@ -11,76 +11,70 @@
 		"build": "tsc && yarn postbuild",
 		"postbuild": "yarn run copy-locales",
 		"copy-locales": "cpy locales/**/*.* dist/ --parents",
-		"start": "node --enable-source-maps --es-module-specifier-resolution=node .",
-		"generate": "VERBOSE=1 DEBUG=1 graphql-codegen --config codegen.yml"
+		"start": "node ."
 	},
 	"imports": {
 		"#util": "./dist/util/index.js",
+		"#util/*": "./dist/util/*.js",
 		"#structs": "./dist/structures/index.js",
 		"#const": "./dist/util/constants.js",
-		"#logger": "./dist/util/logger.js",
-		"#hasura": "./dist/util/hasura.js"
+		"#logger": "./dist/util/logger.js"
 	},
 	"author": "Carter Himmel <fyko@sycer.dev>",
 	"license": "MIT",
 	"dependencies": {
-		"@discordjs/builders": "^0.12.0",
-		"@discordjs/collection": "^0.5.0",
-		"@discordjs/rest": "^0.3.0",
-		"@fyko/export-api": "^0.2.0",
-		"@grpc/grpc-js": "^1.5.7",
+		"@discordjs/builders": "^1.4.0",
+		"@discordjs/collection": "^1.3.0",
+		"@discordjs/rest": "^1.5.0",
+		"@fyko/export-api": "^0.3.1",
+		"@grpc/grpc-js": "^1.8.4",
 		"common-tags": "^1.8.2",
-		"discord.js": "^13.6.0",
-		"express": "^4.17.2",
+		"discord.js": "^14.7.1",
+		"express": "^4.18.2",
 		"fs-nextra": "^0.5.1",
-		"graphql": "^16.3.0",
-		"graphql-request": "^3.7.0",
-		"helmet": "^4.6.0",
-		"i18next": "^21.6.10",
-		"i18next-fs-backend": "^1.1.4",
-		"mongoose": "^6.2.0",
-		"node-fetch": "^2.6.7",
-		"pino": "^6.13.4",
-		"prom-client": "^14.0.1",
+		"helmet": "^6.0.1",
+		"i18next": "^22.4.9",
+		"i18next-fs-backend": "^2.1.1",
+		"node-fetch": "^3.3.0",
+		"pino": "^6.14.0",
+		"postgres": "^3.3.3",
+		"prom-client": "^14.1.1",
 		"reflect-metadata": "^0.1.13",
-		"tslib": "^2.3.1",
-		"tsyringe": "^4.6.0"
+		"tslib": "^2.4.1",
+		"tsyringe": "^4.7.0"
 	},
 	"devDependencies": {
-		"@graphql-codegen/add": "^3.1.1",
-		"@graphql-codegen/cli": "^2.5.0",
-		"@graphql-codegen/typescript": "^2.4.3",
-		"@graphql-codegen/typescript-graphql-request": "^4.3.4",
-		"@graphql-codegen/typescript-operations": "^2.2.4",
 		"@types/common-tags": "^1.8.1",
-		"@types/express": "^4.17.13",
-		"@types/node": "^14.18.10",
-		"@types/node-fetch": "^2.5.12",
+		"@types/express": "^4.17.15",
+		"@types/node": "^18.11.18",
 		"@types/pino": "^6.3.12",
-		"@types/ws": "^7.4.7",
-		"@typescript-eslint/eslint-plugin": "^4.33.0",
-		"@typescript-eslint/parser": "^4.33.0",
-		"cpy-cli": "^3.1.1",
-		"discord-api-types": "^0.26.1",
-		"eslint": "^7.32.0",
-		"eslint-config-marine": "^8.3.2",
-		"eslint-config-prettier": "^7.2.0",
-		"eslint-plugin-prettier": "^3.4.1",
+		"@types/ws": "^8.5.4",
+		"cpy-cli": "^4.2.0",
+		"discord-api-types": "^0.37.28",
+		"eslint": "^8.31.0",
+		"eslint-config-neon": "^0.1.40",
 		"pino-pretty": "^3.6.1",
-		"prettier": "^2.5.1",
-		"rimraf": "^3.0.2",
-		"typescript": "^4.5.5",
-		"winston": "^3.5.1"
+		"prettier": "^2.8.2",
+		"rimraf": "^4.0.4",
+		"typescript": "^4.9.4"
 	},
 	"optionalDependencies": {
-		"bufferutil": "^4.0.6",
+		"bufferutil": "^4.0.7",
 		"erlpack": "github:discord/erlpack",
-		"utf-8-validate": "^5.0.8",
+		"utf-8-validate": "^6.0.0",
 		"zlib-sync": "^0.1.7"
 	},
+	"prettier": {
+		"printWidth": 120,
+		"useTabs": true,
+		"singleQuote": true,
+		"quoteProps": "as-needed",
+		"trailingComma": "all",
+		"endOfLine": "lf"
+	},
 	"volta": {
-		"node": "16.14.0",
+		"node": "18.13.0",
 		"yarn": "1.22.17"
 	},
-	"packageManager": "yarn@3.2.0-rc.11"
+	"packageManager": "yarn@3.3.1"
 }
