@@ -1,27 +1,31 @@
-import { hasuraClient } from '#hasura';
+import { AuditLogEvent } from 'discord-api-types/v9';
+import type { User } from 'discord.js';
+import { Client, EmbedBuilder, Events, PermissionFlagsBits } from 'discord.js';
+import i18next from 'i18next';
+import { Sql } from 'postgres';
+import { inject, injectable } from 'tsyringe';
 import { logger } from '#logger';
 import type { Listener } from '#structs';
-import { Colors } from '#util';
-import { AuditLogEvent } from 'discord-api-types/v9';
-import { Client, Constants, MessageEmbed, Permissions, User } from 'discord.js';
-import i18next from 'i18next';
-import { injectable } from 'tsyringe';
+import { Colors, kSQL } from '#util';
 
 @injectable()
 export default class implements Listener {
-	public readonly event = Constants.Events.GUILD_CREATE;
+	public readonly event = Events.GuildCreate;
 
-	public constructor(public readonly client: Client<true>) {}
+	public constructor(public readonly client: Client<true>, @inject(kSQL) public readonly sql: Sql) {}
 
 	public exec = (): void => {
 		this.client.on(this.event, async (guild) => {
 			logger.info(`[NEW GUILD] Joined a new server! ${guild.name} | ${guild.memberCount} Member(s)`);
-			const existing = await hasuraClient.GetServerByPk({ id: guild.id });
-			if (!existing.archive_utility_server_by_pk) {
-				await hasuraClient.CreateServer({ id: guild.id });
+			const existing = await this.sql`select * from "archive_utility"."server" where id = ${guild.id}`;
+			if (!existing.count) {
+				await this.sql`
+					insert into "archive_utility"."server" (id)
+					values (${guild.id});
+				`;
 			}
 
-			const embed = new MessageEmbed()
+			const embed = new EmbedBuilder()
 				.setAuthor({ name: this.client.user.username, iconURL: this.client.user.displayAvatarURL() })
 				.setDescription(
 					i18next.t('events.guild_create.embeds.invited.description', {
@@ -29,13 +33,13 @@ export default class implements Listener {
 						client_name: this.client.user.username,
 					}),
 				)
-				.setColor(guild.me?.displayColor ?? Colors.Primary);
+				.setColor(guild.members.me?.displayColor ?? Colors.Primary);
 
-			if (guild.me?.permissions.has(Permissions.FLAGS.VIEW_AUDIT_LOG)) {
+			if (guild.members.me?.permissions.has(PermissionFlagsBits.ViewAuditLog)) {
 				const entries = await guild.fetchAuditLogs({ type: AuditLogEvent.BotAdd });
 
 				if (entries.entries.size) {
-					const theEntry = entries.entries.find((e) => (e.target as User).id === this.client.user.id);
+					const theEntry = entries.entries.find((entry) => (entry.target as User).id === this.client.user.id);
 					if (theEntry) {
 						try {
 							await theEntry.executor?.send({ embeds: [embed] });
