@@ -1,29 +1,30 @@
-process.env.NODE_ENV ??= 'development';
 import 'reflect-metadata';
 
-import { Command, Listener, REST } from '#structs';
-import { kCommands, kListeners, kMetrics, kREST, loadCommands, loadListeners, kExportHandler } from '#util';
-import Collection from '@discordjs/collection';
-import { Client, Intents, Options } from 'discord.js';
-import { Gauge, register } from 'prom-client';
-import { container } from 'tsyringe';
+import process from 'node:process';
+import { fileURLToPath, URL } from 'node:url';
+import { Collection } from '@discordjs/collection';
+import { createExporterClient } from '@fyko/export-api/client';
+import { Client, IntentsBitField, Options } from 'discord.js';
 import i18next from 'i18next';
 import Backend from 'i18next-fs-backend';
-import { fileURLToPath, URL } from 'url';
-import { ExporterClient } from '@fyko/export-api/client';
-import { credentials } from '@grpc/grpc-js';
+import postgres from 'postgres';
+import { Gauge, register } from 'prom-client';
+import { container } from 'tsyringe';
+import type { Command, Listener } from '#structs';
+import { kCommands, kExportClient, kListeners, kMetrics, kSQL, loadCommands, loadListeners } from '#util';
+
+process.env.NODE_ENV ??= 'development';
 
 const client = new Client({
-	intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_MESSAGES],
+	intents: [IntentsBitField.Flags.Guilds, IntentsBitField.Flags.GuildMessages],
 	makeCache: Options.cacheWithLimits({
 		MessageManager: 5,
 	}),
 });
-const exportClient = new ExporterClient(process.env.EXPORT_ENDPOINT!, credentials.createInsecure());
 
+const exportClient = createExporterClient(process.env.EXPORT_ENDPOINT!);
 const commands = new Collection<string, Command>();
 const listeners = new Collection<string, Listener>();
-const rest = new REST({ version: '9' }).setToken(process.env.DISCORD_TOKEN!);
 
 const prometheus = {
 	messageCounter: new Gauge({
@@ -33,14 +34,14 @@ const prometheus = {
 	userHistogram: new Gauge({
 		name: 'archive_utility_user_histogram',
 		help: 'Histogram of all users Archive Utility has seen.',
-		collect: function () {
-			return this.set(client.guilds.cache.reduce((acc, g) => (acc += g.memberCount), 0));
+		collect () {
+			this.set(client.guilds.cache.reduce((acc, guild) => (acc + guild.memberCount), 0));
 		},
 	}),
 	guildHistogram: new Gauge({
 		name: 'archive_utility_guild_histogram',
 		help: 'Histogram of Archive Utility Guilds.',
-		collect: function () {
+		collect () {
 			this.set(client.guilds.cache.size);
 		},
 	}),
@@ -52,17 +53,23 @@ const prometheus = {
 };
 export type Metrics = typeof prometheus;
 
+const sql = postgres(process.env.DATABASE_URL!, {
+	types: {
+		bigint: postgres.BigInt,
+	},
+});
+
 container.register(Client, { useValue: client });
 container.register(kCommands, { useValue: commands });
 container.register(kListeners, { useValue: listeners });
-container.register(kREST, { useValue: rest });
 container.register(kMetrics, { useValue: prometheus });
-container.register(kExportHandler, { useValue: exportClient });
+container.register(kExportClient, { useValue: exportClient });
+container.register(kSQL, { useValue: sql });
 
 async function start() {
 	await i18next.use(Backend).init({
 		backend: {
-			loadPath: fileURLToPath(new URL('./locales/{{lng}}/{{ns}}.json', import.meta.url)),
+			loadPath: fileURLToPath(new URL('locales/{{lng}}/{{ns}}.json', import.meta.url)),
 		},
 		cleanCode: true,
 		fallbackLng: ['en-US'],
