@@ -1,6 +1,8 @@
-import { logger } from '#logger';
+import process from 'node:process';
 import { REST } from '@discordjs/rest';
-import { RESTPutAPIApplicationCommandsResult, Routes } from 'discord-api-types/v9';
+import type { RESTPutAPIApplicationCommandsResult} from 'discord-api-types/v9';
+import { Routes } from 'discord-api-types/v9';
+import { logger } from '#logger';
 
 const rest = new REST({ version: '9' }).setToken(process.env.DISCORD_TOKEN!);
 
