@@ -1,7 +1,20 @@
-export const kExportHandler = Symbol('ExportHandler');
-export const kSettings = Symbol('Settings');
-export const kREST = Symbol('REST');
+/**
+ * The gRPC client for the export service.
+ */
+export const kExportClient = Symbol('ExportClient');
+/**
+ * A collection of commands.
+ */
 export const kCommands = Symbol('Commands');
+/**
+ * A collection of listeners.
+ */
 export const kListeners = Symbol('Listeners');
-export const kClient = Symbol('Client');
+/**
+ * The Prometheus metrics.
+ */
 export const kMetrics = Symbol('Metrics');
+/**
+ * The SQL connection.
+ */
+export const kSQL = Symbol('SQL');
